refactor: 대각선이 아닌 꼭짓점을 나타내는 표현 변경
